requirements.txt
torch>=2.2
tqdm
pyyaml
